9th commit: program restructured
