+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(foodOrdering C)
 
 set(CMAKE_C_STANDARD 99)
 
-add_executable(foodOrdering main.c)
+add_executable(foodOrdering main.c menu.h menu.c user.h order.h user.c choice.h choice.c order.c)
