+++ choice.h
@@ -1,0 +1,13 @@
+//
+// Created by Bori on 11/3/2019.
+//
+
+#ifndef FOODORDERING_CHOICE_H
+#define FOODORDERING_CHOICE_H
+
+#include <stdbool.h>
+
+bool correctAnswer(char choice, char min, char max);
+int getChoiceIndex(int nrOfChoices);
+
+#endif //FOODORDERING_CHOICE_H
